chore: rename to UserDropdown.tsx
