+++ site/src/components/Navbar/NavbarView.tsx
@@ -11,7 +11,7 @@ import { colors } from "theme/colors"
 import * as TypesGen from "../../api/typesGenerated"
 import { navHeight } from "../../theme/constants"
 import { combineClasses } from "../../utils/combineClasses"
-import { UserDropdown } from "../UserDropdown/UsersDropdown"
+import { UserDropdown } from "../UserDropdown/UserDropdown"
 import Box from "@mui/material/Box"
 import Menu from "@mui/material/Menu"
 import Button from "@mui/material/Button"
