+++ site/src/components/UserDropdown/UserDropdown.stories.tsx
@@ -1,7 +1,7 @@
 import Box from "@mui/material/Box"
 import { Story } from "@storybook/react"
 import { MockUser } from "../../testHelpers/entities"
-import { UserDropdown, UserDropdownProps } from "./UsersDropdown"
+import { UserDropdown, UserDropdownProps } from "./UserDropdown"
 
 export default {
   title: "components/UserDropdown",
