+++ site/src/components/UserDropdown/UserDropdown.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, screen } from "@testing-library/react"
 import { MockSupportLinks, MockUser } from "../../testHelpers/entities"
 import { render } from "../../testHelpers/renderHelpers"
 import { Language } from "../UserDropdownContent/UserDropdownContent"
-import { UserDropdown, UserDropdownProps } from "./UsersDropdown"
+import { UserDropdown, UserDropdownProps } from "./UserDropdown"
 
 const renderAndClick = async (props: Partial<UserDropdownProps> = {}) => {
   render(
